refactor: rename Tailwind setup module
- Renamed the Tailwind setup function module from `tailwind.ts` to `tailwind-setup.js`.
