+++ cli/src/commands/add/tailwind-setup.ts
